+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pre-release
 
 ### Implemented enhancements
+- CI-439 - Added `ExpandInternationalLandingPage` new FDI landing page
 
 ### Fixed Bugs